+++ core/pkg/artifacts/saver.go
@@ -39,6 +39,7 @@ type ArtifactSaver struct {
 }
 
 type multipartUploadInfo = []gql.CreateArtifactFilesCreateArtifactFilesCreateArtifactFilesPayloadFilesFileConnectionEdgesFileEdgeNodeFileUploadMultipartUrlsUploadUrlPartsUploadUrlPart
+type updateArtifactManifestAttrs = gql.UpdateArtifactManifestUpdateArtifactManifestUpdateArtifactManifestPayloadArtifactManifest
 
 type serverFileResponse struct {
 	name            string
@@ -152,6 +153,44 @@ func (as *ArtifactSaver) createManifest(
 		return gql.CreateArtifactManifestCreateArtifactManifestCreateArtifactManifestPayloadArtifactManifest{}, err
 	}
 	return response.GetCreateArtifactManifest().ArtifactManifest, nil
+}
+
+func (as *ArtifactSaver) updateManifest(
+	artifactManifestId string, manifestDigest string,
+) (attrs updateArtifactManifestAttrs, rerr error) {
+	response, err := gql.UpdateArtifactManifest(
+		as.Ctx,
+		as.GraphqlClient,
+		artifactManifestId,
+		&manifestDigest,
+		nil,
+		true,
+	)
+	if err != nil {
+		return updateArtifactManifestAttrs{}, err
+	}
+	if response == nil || response.GetUpdateArtifactManifest() == nil {
+		return updateArtifactManifestAttrs{}, fmt.Errorf("received invalid response from UpdateArtifactManifest")
+	}
+	return response.GetUpdateArtifactManifest().ArtifactManifest, nil
+}
+
+func (as *ArtifactSaver) upsertManifest(
+	artifactId string, baseArtifactId *string, artifactManifestId string, manifestDigest string,
+) (uploadUrl *string, uploadHeaders []string, rerr error) {
+	if as.Artifact.IncrementalBeta1 || as.Artifact.DistributedId != "" {
+		updateManifestAttrs, err := as.updateManifest(artifactManifestId, manifestDigest)
+		if err != nil {
+			return nil, nil, fmt.Errorf("ArtifactSaver.updateManifest: %w", err)
+		}
+		return updateManifestAttrs.File.UploadUrl, updateManifestAttrs.File.UploadHeaders, nil
+	} else {
+		manifestAttrs, err := as.createManifest(artifactId, baseArtifactId, manifestDigest, true /* includeUpload */)
+		if err != nil {
+			return nil, nil, fmt.Errorf("ArtifactSaver.createManifest: %w", err)
+		}
+		return manifestAttrs.File.UploadUrl, manifestAttrs.File.UploadHeaders, nil
+	}
 }
 
 func (as *ArtifactSaver) uploadFiles(
@@ -647,11 +686,13 @@ func (as *ArtifactSaver) Save(ch chan<- *service.Record) (artifactID string, rer
 		return "", fmt.Errorf("ArtifactSaver.writeManifest: %w", err)
 	}
 	defer os.Remove(manifestFile)
-	manifestAttrs, err = as.createManifest(artifactID, baseArtifactId, manifestDigest, true /* includeUpload */)
+
+	uploadUrl, uploadHeaders, err := as.upsertManifest(artifactID, baseArtifactId, manifestAttrs.Id, manifestDigest)
 	if err != nil {
-		return "", fmt.Errorf("ArtifactSaver.createManifest: %w", err)
+		return "", fmt.Errorf("ArtifactSaver.upsertManifest: %w", err)
 	}
-	err = as.uploadManifest(manifestFile, manifestAttrs.File.UploadUrl, manifestAttrs.File.UploadHeaders, ch)
+
+	err = as.uploadManifest(manifestFile, uploadUrl, uploadHeaders, ch)
 	if err != nil {
 		return "", fmt.Errorf("ArtifactSaver.uploadManifest: %w", err)
 	}